feat(audit): Pass project to WeekCard for context
The `Project` object was not being passed down to the `WeekCard` component. This commit passes the `project` prop to `WeekCard` so that it can be used within the component. This change also updates the `WeekCard` and `EventItem` components to use the newly available `project` prop and cleans up unused imports.